RAG_OwnFile_pdf-summarizer/requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-google-genai
pypdf
faiss-cpu
